+++ src/providers/openai/moderation.ts
@@ -1,0 +1,108 @@
+import type { Cache } from 'cache-manager';
+import OpenAI from 'openai';
+import { OpenAiGenericProvider } from '.';
+import { getCache, isCacheEnabled } from '../../cache';
+import logger from '../../logger';
+import type {
+  ApiModerationProvider,
+  ModerationFlag,
+  ProviderModerationResponse,
+} from '../../types';
+import { REQUEST_TIMEOUT_MS } from '../shared';
+
+export class OpenAiModerationProvider
+  extends OpenAiGenericProvider
+  implements ApiModerationProvider
+{
+  async callModerationApi(
+    userPrompt: string, // userPrompt is not supported by OpenAI moderation API
+    assistantResponse: string,
+  ): Promise<ProviderModerationResponse> {
+    if (!this.getApiKey()) {
+      throw new Error(
+        'OpenAI API key is not set. Set the OPENAI_API_KEY environment variable or add `apiKey` to the provider config.',
+      );
+    }
+
+    const openai = new OpenAI({
+      apiKey: this.getApiKey(),
+      organization: this.getOrganization(),
+      baseURL: this.getApiUrl(),
+      maxRetries: 3,
+      timeout: REQUEST_TIMEOUT_MS,
+    });
+
+    let cache: Cache | undefined;
+    let cacheKey: string | undefined;
+    if (isCacheEnabled()) {
+      cache = await getCache();
+      cacheKey = `openai:${this.modelName}:${JSON.stringify(
+        this.config,
+      )}:${userPrompt}:${assistantResponse}`;
+
+      // Try to get the cached response
+      const cachedResponse = await cache.get(cacheKey);
+
+      if (cachedResponse) {
+        logger.debug(`Returning cached response for ${userPrompt}: ${cachedResponse}`);
+        return JSON.parse(cachedResponse as string);
+      }
+    }
+
+    logger.debug(
+      `Calling OpenAI moderation API: prompt [${userPrompt}] assistant [${assistantResponse}]`,
+    );
+    let moderation: OpenAI.Moderations.ModerationCreateResponse | undefined;
+    try {
+      moderation = await openai.moderations.create({
+        model: this.modelName,
+        input: assistantResponse,
+      });
+    } catch (err) {
+      logger.error(`API call error: ${String(err)}`);
+      return {
+        error: `API call error: ${String(err)}`,
+      };
+    }
+
+    logger.debug(`\tOpenAI moderation API response: ${JSON.stringify(moderation)}`);
+    try {
+      const { results } = moderation;
+
+      const flags: ModerationFlag[] = [];
+      if (!results) {
+        throw new Error('API response error: no results');
+      }
+
+      if (cache && cacheKey) {
+        await cache.set(cacheKey, JSON.stringify(moderation));
+      }
+
+      if (results.length === 0) {
+        return { flags };
+      }
+
+      for (const result of results) {
+        if (result.flagged) {
+          for (const [category, flagged] of Object.entries(result.categories)) {
+            if (flagged) {
+              flags.push({
+                code: category,
+                description: category,
+                confidence:
+                  result.category_scores[category as keyof OpenAI.Moderation.CategoryScores],
+              });
+            }
+          }
+        }
+      }
+      return { flags };
+    } catch (err) {
+      return {
+        error: `API response error: ${String(err)}: ${JSON.stringify(moderation)}`,
+      };
+    }
+  }
+}
+
+export const DefaultModerationProvider = new OpenAiModerationProvider('omni-moderation-latest');
